backend/requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.6
django-cors-headers==3.8.0
django-samesite-none==0.0.3
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
PyJWT==2.3.0
python-decouple==3.4
pytz==2021.3
sqlparse==0.4.2
static3==0.7.0
djangorestframework_simplejwt==4.7.2
